LayoutTests/imported/w3c/web-platform-tests/content-security-policy/inheritance/blob-url-inherits-from-initiator.sub-expected.txt


PASS Initiator is same-origin with target frame.
PASS Initiator is cross-origin with target frame.
